pyproject.toml
[tool.poetry]
name = "rtaa-classifier"
version = "1.0.0"
description = "A real-time Twitter Classifier"
authors = ["andefined <dimitris.papaevagelou@andefined.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.48.0"
catboost = "^1.0.6"
numpy = "^1.23.2"
python-dotenv = "^0.20.0"
transformers = "^4.21.2"
textacy = "^0.12.0"
emoji = "^2.0.0"
demoji = "^1.1.0"
grpcio-tools = "^1.48.0"
Unidecode = "^1.3.4"
torch = "^1.12.1"
grpcio-status = "^1.48.0"
detoxify = "^0.5.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
